Fix compilation errors in payment-service
Fixes two compilation errors identified in CI build:

1. PaymentGatewayService.java - Type conversion error
   - Convert BigDecimal to Double using doubleValue()
   - Payment.amount is BigDecimal but PaymentGatewayRequest expects Double
   - Fixed at line 31: .amount(payment.getAmount().doubleValue())

2. RabbitMQConfig.java - Missing routing key constants
   - Add DONATION_CAPTURED_ROUTING_KEY = "donation.captured"
   - Add DONATION_AUTHORIZED_ROUTING_KEY = "donation.authorized"
   - Add DONATION_FAILED_ROUTING_KEY = "donation.failed"
   - These constants are referenced by OutboxPublisher.java

These fixes resolve all compilation errors and the build should now pass.

+++ payment-service/src/main/java/com/careforall/payment/service/PaymentGatewayService.java
@@ -28,7 +28,7 @@ public class PaymentGatewayService {
         try {
             PaymentGatewayRequest request = PaymentGatewayRequest.builder()
                     .pledgeId(payment.getPledgeId())
-                    .amount(payment.getAmount())
+                    .amount(payment.getAmount().doubleValue())
                     .campaignId(payment.getCampaignId())
                     .userId(payment.getUserId())
                     .build();
